add hatch command
