+++ pyproject.toml
@@ -288,6 +288,9 @@ dependencies = [
 [tool.hatch.envs.gql.scripts]
 build = 'strawberry export-schema phoenix.server.api.schema:schema > app/schema.graphql'
 
+[tool.hatch.envs.openapi.scripts]
+build = 'python -m phoenix.server.api.openapi.export > openapi.json'
+
 [tool.hatch.envs.proto]
 detached = true
 dependencies = [
