+++ src/phoenix/server/app.py
@@ -72,7 +72,7 @@ from phoenix.server.api.dataloaders import (
     TokenCountDataLoader,
     TraceEvaluationsDataLoader,
 )
-from phoenix.server.api.openapi import OPENAPI_SCHEMA_GENERATOR
+from phoenix.server.api.openapi.schema import OPENAPI_SCHEMA_GENERATOR
 from phoenix.server.api.routers.v1 import V1_ROUTES
 from phoenix.server.api.schema import schema
 from phoenix.server.grpc_server import GrpcServer
